Create Configuring the optimizer
Configuring the optimizer
Now that we have training logic, we need to specify how to optimize the model's parameters.

In this exercise, you'll complete the configure_optimizers method within a PyTorch Lightning module used for image classification tasks. Your goal is to set up an optimizer that will update the model's parameters during training. To complete this you'll use the Adam optimizer with a learning rate of 1e-3.

Instructions

Create an Adam optimizer using the model's parameters, setting the learning rate to 1e-3.