Create Training and evaluating
Training and evaluating
In this exercise, we'll bring together al that we've practiced so far, by training and evaluating a neural network on the real-world dataset of handwritten Ethiopian MNIST characters.

ImageClassifier is a predefined neural network model implemented using PyTorch Lightning. It consists of convolutional layers for feature extraction, activation functions for introducing non-linearity, and fully connected layers for classification.

The Ethiopic MNIST dataset was pre-imported for you.

Instructions

Import the Trainer.
Define ImageClassifier model and trainer.
Train the model.
Evaluate the model on the validation set.